refactor: 댓글 관련 로직 위치 변경 (processor-service -> user-service)
