+++ user-service/src/main/java/com/example/devnote/config/PasswordConfig.java
@@ -1,13 +1,10 @@
-package com.example.devnote.processor_service.config;
+package com.example.devnote.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-/**
- * 비밀번호 해시용 PasswordEncoder 빈 등록
- */
 @Configuration
 public class PasswordConfig {
     @Bean
