+++ user-service/src/main/java/com/example/devnote/dto/CommentRequestDto.java
@@ -1,18 +1,14 @@
-package com.example.devnote.processor_service.dto;
+package com.example.devnote.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-/**
- * 댓글 요청 dto
- */
 @Data
 public class CommentRequestDto {
     @NotNull(message = "contentId is required")
     private Long contentId;
 
-    /** 대댓글이라면 parentId 지정 */
     private Long parentId;
 
     @NotBlank(message = "username is required")
