+++ user-service/src/main/java/com/example/devnote/dto/CommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.devnote.processor_service.dto;
+package com.example.devnote.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,22 +8,17 @@ import lombok.NoArgsConstructor;
 import java.time.Instant;
 import java.util.List;
 
-/**
- * 댓글 반환 DTO
- */
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class CommentResponseDto {
     private Long id;
-    private Long parentId; // null 이면 루트 댓글
+    private Long parentId;
     private Long contentId;
     private String username;
     private String content;
     private Instant createdAt;
     private Instant updatedAt;
-
-    /** 답글이 있을 경우 포함 */
     private List<CommentResponseDto> replies;
 }
